requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow-cpu
transformers
datasets
xgboost
mlflow
nltk
Flask
boto3
python-dotenv